+++ src/containers/BookList.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import Book from './Book';
+import Book from '../components/Book';
 import { removeBook } from '../actions/index';
 
 class BookList extends Component {
